requirements.txt
asgiref==3.8.1
attrs==24.2.0
coverage==7.6.1
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
